Ambiente pruebas (#37)
* Simv5 busqueda documentos rey004 (#29)

* sdff

* Ajuste buscar por indices Abril 20 2023

* Ajuste Abril 20 2023 Noche

* dads

* sfdsf

* Ajuste indices Abril 23 2023

---------


* Si mv5 calculo evaluacion val tra rey 005 (#30)

* Ajuste hoy Calculo

* x

* Ajuste Abril 17 2023

* Corrige N_Realcion nulo Abril 23 2023

* Ajuste logo Abril 23 2023

---------


* Ajuste Cache de la layout

* Reposiciones flora (#31)

* Se realizan todos los ajustes solicitados en el formulario de reposiciones

* Mejoras

* Se realizan las mejoras solicitadas al reporte de seguimiento a trámites nuevos

* Se realizan ajustes finales

* rr

* Si mv5 plantilla control rene0001 (#33)

* Control y Vigilancia: Ajuste en la carga de algunos índices del documento de Word

* Facturación de Informes Técnicos: Colocar filtros en todas las columnas, Eliminar la columna PARA.
Agregar Columna fecha radicado, Filtrar por Entidad desde la fuente de los datos.

* tyt

---------


* dsfs

* Expedientes jorge 007 (#34)

* Ambiente pruebas (#32)

* Simv5 busqueda documentos rey004 (#29)

* sdff

* Ajuste buscar por indices Abril 20 2023

* Ajuste Abril 20 2023 Noche

* dads

* sfdsf

* Ajuste indices Abril 23 2023

---------


* Si mv5 calculo evaluacion val tra rey 005 (#30)

* Ajuste hoy Calculo

* x

* Ajuste Abril 17 2023

* Corrige N_Realcion nulo Abril 23 2023

* Ajuste logo Abril 23 2023

---------


* Ajuste Cache de la layout

* Reposiciones flora (#31)

* Se realizan todos los ajustes solicitados en el formulario de reposiciones

* Mejoras

* Se realizan las mejoras solicitadas al reporte de seguimiento a trámites nuevos

* Se realizan ajustes finales

* rr

---------


* sada

* adasad

* s

* Se agrega la seguridad

---------


* Reposiciones flora (#36)

* Se realizan todos los ajustes solicitados en el formulario de reposiciones

* Mejoras

* Se realizan las mejoras solicitadas al reporte de seguimiento a trámites nuevos

* Se realizan ajustes finales

* Se ajustan los 6 digitos decimales para laso coordenadas x y y de los dos formularios

* Mejoras

---------

+++ SIMv5/Areas/ControlVigilancia/Models/Reposicion.cs
@@ -81,10 +81,10 @@
         public string Asunto { get; set; }
 
         [JsonProperty("coordenadaX")]
-        public float? CoordenadaX { get; set; }
+        public double? CoordenadaX { get; set; }
 
         [JsonProperty("coordenadaY")]
-        public float? CoordenadaY { get; set; }
+        public double? CoordenadaY { get; set; }
 
         [JsonProperty("nroLeniosSolicitados")]
         public decimal? NroLeniosSolicitados { get; set; }
